+++ CHANGELOG.rst
@@ -2,6 +2,10 @@
 Changelog for package rodney
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
+Out standing changes
+------------------
+* Moved the scaning for faces action client to rondey_missions package
+
 0.2.0 (2018-06-24)
 ------------------
 * Moved Arduino sketch into package